ERestaurant Management System
(Web Application using Servlet and JSP)   The system aims to provide
online ordering and reservation services to the customers. The customers
visit the  ordering webpage where they are presented with an interactive
and up-to-date menu with all available options and  dynamically
adjusting prices based on the selected options. The customers can also
choose and book available tables.  Further there is a manager’s access
who can change the menu and their prices and also approve and mark
delivery of  customer’s orders.